requirements.txt
# Core Framework
streamlit==1.28.1
flask==2.3.3
flask-cors==4.0.0
flask-restx==1.2.0
gunicorn==21.2.0

# AI & ML
openai==1.3.8
anthropic==0.7.8
tiktoken==0.5.2

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# File Processing
pypdf2==3.0.1
python-docx==1.1.0
pillow>=10.2.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
requests==2.31.0
pandas==2.1.4
plotly==5.17.0

# Security
bcrypt==4.1.2
pyjwt==2.8.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1